+++ app/src/main/java/com/example/symptotrack/net/dto/UserData.java
@@ -1,10 +1,10 @@
 package com.example.symptotrack.net.dto;
 
 public class UserData {
-    public int id;
+    public long id;
     public String first_name;
     public String last_name;
     public String phone;
-    public String email;
-    public String username;
+    public String email;    // puede venir null si se registró con username
+    public String username; // puede venir null si se registró con email
 }
